change the Id type to long to match the type size from the db

+++ ZawodyWin/ZawodyWin/DataModels/Competition.cs
@@ -2,9 +2,9 @@
 {
     public class Competition
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public int SeriesNumber { get; set; }
-        public int TournametId { get; set; }
+        public long TournametId { get; set; }
     }
 }
